requirements.txt
gradio==6.0.2
matplotlib
Pillow